Move AuthService into separate services directory
